MAINT: Separate alertdialog from main app.js
