+++ Procfile
@@ -1,1 +1,1 @@
-web: uvicorn --host 0.0.0.0 --port 8000 src.random_dish.main:app
+web: poetry run random_dish
